add mirgecom/array_context.py, refactor
