+++ doc/support/tools.rst
@@ -3,3 +3,4 @@ Random Pile'o'Tools
 
 .. automodule:: mirgecom.simutil
 .. automodule:: mirgecom.utils
+.. automodule:: mirgecom.array_context
